fleshed out testing some more. completed unit supply interogation and depletion.

+++ GUBS.Components/ActionDetractor.cpp
@@ -5,7 +5,7 @@
 
 namespace GUBS_Supply
 {
-	ActionDetractor::ActionDetractor(float moveDetractor, float attDetractor, float defDetractor)
+	ActionDetractor::ActionDetractor(double moveDetractor, double attDetractor, double defDetractor)
 		: _Movement(moveDetractor), _Attack(attDetractor), _Defense(defDetractor)
 	{
