+++ GUBS.Components/Supply/SupplyTypes/Supply_Lubrication.h
@@ -25,7 +25,7 @@ namespace GUBS_Supply
 			_Type = SupplyType::LUBRICATION;
 		}
 
-		Supply_Lubrication(unsigned long id, const std::string& supplyName, const std::string& description, SupplySubType subType, MeasurementUnit unit, float massPer, const Volume& volumePer)
+		Supply_Lubrication(unsigned long id, const std::string& supplyName, const std::string& description, SupplySubType subType, MeasurementUnit unit, double massPer, const Volume& volumePer)
 			:_Supply(id, supplyName, description, SupplyType::LUBRICATION, subType, unit, massPer, volumePer)
 		{
 			DBUG("Supply_Lubrication");
